==> model/2-credits_preprocessing.ipynb <==
# Importing Modules

import numpy as np
import pandas as pd

## Read File
- Dataset can be download from [credits.csv](https://www.kaggle.com/datasets/rounakbanik/the-movies-dataset?select=credits.csv)

credits = pd.read_csv('../datasets/credits.csv')

credits

## Read Also

- Dataset can be download from [movies_metadata.csv](https://www.kaggle.com/datasets/rounakbanik/the-movies-dataset?select=movies_metadata.csv)

# meta = pd.read_csv('../datasets/movies_metadata.csv')
p_file = '../datasets/movies_metadata.csv'
meta = pd.read_csv(p_file, low_memory=False)

meta.head()

meta['release_date'] = pd.to_datetime(meta['release_date'], errors='coerce')

meta['year'] = meta['release_date'].dt.year

meta['year'].value_counts().sort_index()

**Note:** As you can see data are present from year 1874 to 2020.

# Filtering

- Getting only 2017 movies as we already have movies up to the year 2016 in preprocessing 1 file. 
- We don't have enough data for the movies from 2018, 2019 and 2020. 
- We'll deal with it in the upcoming preprocessing files

new_meta = meta.loc[meta.year == 2017,['genres','id','title','year']]

new_meta

new_meta['id'] = new_meta['id'].astype(int)

data = pd.merge(new_meta, credits, on='id')

pd.set_option('display.max_colwidth', 75)

data

## Replacing 

evaluates an expression node or a string containing a Python literal or container display

import ast
data['genres'] = data['genres'].map(lambda x: ast.literal_eval(x))
data['cast'] = data['cast'].map(lambda x: ast.literal_eval(x))
data['crew'] = data['crew'].map(lambda x: ast.literal_eval(x))

def make_genresList(x):
    gen = []
    st = " "
    for i in x:
        if i.get('name') == 'Science Fiction':
            scifi = 'Sci-Fi'
            gen.append(scifi)
        else:
            gen.append(i.get('name'))
    if gen == []:
        return np.NaN
    else:
        return (st.join(gen))

data['genres_list'] = data['genres'].map(lambda x: make_genresList(x))

data['genres_list']

def get_actor1(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
    if casts == []:
        return np.NaN
    else:
        return (casts[0])

data['actor_1_name'] = data['cast'].map(lambda x: get_actor1(x))

def get_actor2(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
    if casts == [] or len(casts)<=1:
        return np.NaN
    else:
        return (casts[1])

data['actor_2_name'] = data['cast'].map(lambda x: get_actor2(x))

data['actor_2_name']

def get_actor3(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
    if casts == [] or len(casts)<=2:
        return np.NaN
    else:
        return (casts[2])

data['actor_3_name'] = data['cast'].map(lambda x: get_actor3(x))

data['actor_3_name']

def get_directors(x):
    dt = []
    st = " "
    for i in x:
        if i.get('job') == 'Director':
            dt.append(i.get('name'))
    if dt == []:
        return np.NaN
    else:
        return (st.join(dt))

data['director_name'] = data['crew'].map(lambda x: get_directors(x))

data['director_name']

movie = data.loc[:,['director_name','actor_1_name','actor_2_name','actor_3_name','genres_list','title']]

movie

movie.isna().sum()

movie = movie.dropna(how='any')

movie.isna().sum()

movie = movie.rename(columns={'genres_list':'genres'})
movie = movie.rename(columns={'title':'movie_title'})

movie['movie_title'] = movie['movie_title'].str.lower()

movie['comb'] = movie['actor_1_name'] + ' ' + movie['actor_2_name'] + ' '+ movie['actor_3_name'] + ' '+ movie['director_name'] +' ' + movie['genres']

movie

old = pd.read_csv('new_movie_metadata.csv')

old

old['comb'] = old['actor_1_name'] + ' ' + old['actor_2_name'] + ' '+ old['actor_3_name'] + ' '+ old['director_name'] +' ' + old['genres']

old

# append is deprecated by pandas
new = old._append(movie)
# new = old.concat(movie)

new

new.drop_duplicates(subset ="movie_title", keep = 'last', inplace = True)

new

new.to_csv('new_credits.csv',index=False)
